Move file testing utilities to conftest.
