+++ pyproject.toml
@@ -61,5 +61,5 @@ disable = """
     missing-module-docstring,
 """
 ignore-patterns = "^_.*" # Ignore files that start with an underscore, i.e. _version.py
-
+extension-pkg-allow-list="pyarrow.lib"
