+++ tests/hipscat_import/test_map_reduce.py
@@ -3,7 +3,6 @@
 import os
 import tempfile
 
-import file_testing as ft
 import hipscat.pixel_math as hist
 import numpy.testing as npt
 import pyarrow as pa
@@ -98,7 +97,7 @@ def test_map_headers_wrong(formats_headers_csv):
             )
 
 
-def test_map_headers(formats_headers_csv):
+def test_map_headers(formats_headers_csv, assert_parquet_file_ids):
     """Test loading the a file with non-default headers"""
     with tempfile.TemporaryDirectory() as tmp_dir:
         result = mr.map_to_pixels(
@@ -120,13 +119,13 @@ def test_map_headers(formats_headers_csv):
 
         file_name = os.path.join(tmp_dir, "pixel_11", "shard_0_0.parquet")
         expected_ids = [*range(700, 708)]
-        ft.assert_parquet_file_ids(file_name, "object_id", expected_ids)
+        assert_parquet_file_ids(file_name, "object_id", expected_ids)
 
         file_name = os.path.join(tmp_dir, "pixel_1", "shard_0_0.parquet")
         assert not os.path.exists(file_name)
 
 
-def test_map_small_sky_order0(small_sky_single_file):
+def test_map_small_sky_order0(small_sky_single_file, assert_parquet_file_ids):
     """Test loading the small sky catalog and partitioning each object into the same large bucket"""
     with tempfile.TemporaryDirectory() as tmp_dir:
         result = mr.map_to_pixels(
@@ -148,10 +147,10 @@ def test_map_small_sky_order0(small_sky_single_file):
 
         file_name = os.path.join(tmp_dir, "pixel_11", "shard_0_0.parquet")
         expected_ids = [*range(700, 831)]
-        ft.assert_parquet_file_ids(file_name, "id", expected_ids)
+        assert_parquet_file_ids(file_name, "id", expected_ids)
 
 
-def test_map_small_sky_part_order1(small_sky_file0):
+def test_map_small_sky_part_order1(small_sky_file0, assert_parquet_file_ids):
     """
     Test loading a small portion of the small sky catalog and
     partitioning objects into four smaller buckets
@@ -178,25 +177,25 @@ def test_map_small_sky_part_order1(small_sky_file0):
         # Pixel 44 - contains 5 objects
         file_name = os.path.join(tmp_dir, "pixel_44", "shard_0_0.parquet")
         expected_ids = [703, 707, 716, 718, 723]
-        ft.assert_parquet_file_ids(file_name, "id", expected_ids)
+        assert_parquet_file_ids(file_name, "id", expected_ids)
 
         # Pixel 45 - contains 7 objects
         file_name = os.path.join(tmp_dir, "pixel_45", "shard_0_0.parquet")
         expected_ids = [704, 705, 710, 719, 720, 722, 724]
-        ft.assert_parquet_file_ids(file_name, "id", expected_ids)
+        assert_parquet_file_ids(file_name, "id", expected_ids)
 
         # Pixel 46 - contains 11 objects
         file_name = os.path.join(tmp_dir, "pixel_46", "shard_0_0.parquet")
         expected_ids = [700, 701, 706, 708, 709, 711, 712, 713, 714, 715, 717]
-        ft.assert_parquet_file_ids(file_name, "id", expected_ids)
+        assert_parquet_file_ids(file_name, "id", expected_ids)
 
         # Pixel 47 - contains 2 objects
         file_name = os.path.join(tmp_dir, "pixel_47", "shard_0_0.parquet")
         expected_ids = [702, 721]
-        ft.assert_parquet_file_ids(file_name, "id", expected_ids)
+        assert_parquet_file_ids(file_name, "id", expected_ids)
 
 
-def test_reduce_order0(parquet_shards_dir):
+def test_reduce_order0(parquet_shards_dir, assert_parquet_file_ids):
     """Test reducing into one large pixel"""
     with tempfile.TemporaryDirectory() as tmp_dir:
         mr.reduce_pixel_shards(
@@ -212,7 +211,7 @@ def test_reduce_order0(parquet_shards_dir):
         output_file = os.path.join(tmp_dir, "Norder0/Npix11", "catalog.parquet")
 
         expected_ids = [*range(700, 831)]
-        ft.assert_parquet_file_ids(output_file, "id", expected_ids)
+        assert_parquet_file_ids(output_file, "id", expected_ids)
 
 
 def test_reduce_bad_expectation(parquet_shards_dir):
